GA_denoise_selector/requirements.txt
numpy>=2.0
pandas>=2.0
matplotlib>=3.8
pygad>=3.3
scipy>=1.11
scikit-image>=0.25
scikit-learn>=1.5
PyWavelets>=1.4
PyYAML>=6.0
